voice_agent_project/requirements.txt
django
fastapi
uvicorn
prisma
deepgram-sdk
twilio
python-dotenv
transformers
torch
bentoml
tenacity
aiohttp
cachetools
prometheus-client
psycopg2
tiktoken 
protobuf
setuptools
requests
PyPDF2
datasets
scikit-learn
huggingface_hub[hf_xet]
transformers[torch]